consumer offset truncate hook
